// src/graphql/generated/graphql.tsx
export type ReservationFragment = {
  createdAt: any;
  email: string;
  endDateTime: any;
  id: number;
  noOfAttendees: number;
  paymentStatus: string;
  pricePerPersonInDollars: number;
  slotId: number;
  startDateTime: any;
  title: string;
  updatedAt: any;
  username?: string | null | undefined;
  wineryName: string;
  getUrl?: string | null | undefined;
};